requirements.txt
pyserial
simpleaudio